+++ src/DMX/Fixture/Parameters/DimmerParameter.h
@@ -14,7 +14,7 @@ namespace DMX::Parameters
     {
         static constexpr std::size_t k_size = sizeof...(TTypes) * TBytes;
 
-        DimmerParameter(uint8_t* bytes, uint16_t baseOffset)
+        DimmerParameter(uint8_t* bytes, uint32_t baseOffset)
             : Parameter(bytes, baseOffset, TBytes, ParameterTypes::DIMMER, k_size)
         {
             int16_t index = -TBytes;
