+++ src/DMX/Fixture/Parameters/PositionParameter.h
@@ -16,7 +16,7 @@ namespace DMX::Parameters
     {
         static constexpr std::size_t k_size = sizeof...(TTypes) * TBytes;
 
-        PositionParameter(uint8_t* bytes, uint16_t baseOffset)
+        PositionParameter(uint8_t* bytes, uint32_t baseOffset)
             : Parameter(bytes, baseOffset, TBytes, ParameterTypes::POSITION, k_size)
         {
             int16_t index = -TBytes;
